+++ src/extension_consts.ml
@@ -70,6 +70,8 @@ module Commands = struct
   let copy_type_under_cursor = ocaml_prefixed "copy-type-under-cursor"
 
   let construct = ocaml_prefixed "construct"
+
+  let merlin_jump = ocaml_prefixed "jump"
 end
 
 module Command_errors = struct